pyproject.toml
[tool.poetry]
name = "pyvcard"
version = "0.1.2"
description = ""
authors = ["rodrigo-fernandez <rodrigof@epidataconsulting.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.0.1"
qrcode = "^7.3.1"
vobject = "^0.9.6"
segno = "^1.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
